toolkit/components/url-classifier/tests/mochitest/classifierFrame.html

<html>
<head>
<title></title>

<script type="text/javascript">

var scriptItem = "untouched";

function checkLoads() {
  // Make sure the javascript did not load.
  window.parent.is(scriptItem, "untouched", "Should not load bad javascript");

  // Make sure the css did not load.
  var elt = document.getElementById("styleCheck");
  var style = document.defaultView.getComputedStyle(elt, "");
  window.parent.isnot(style.visibility, "hidden", "Should not load bad css");

  // End (parent) test.
  window.parent.SimpleTest.finish();
}

</script>

<!-- Try loading from a malware javascript URI -->
<script type="text/javascript" src="http://malware.example.com/tests/toolkit/components/url-classifier/tests/mochitest/evil.js"></script>

<!-- Try loading from a malware css URI -->
<link rel="stylesheet" type="text/css" href="http://malware.example.com/tests/toolkit/components/url-classifier/tests/mochitest/evil.css"></link>

<!-- XXX How is this part of the test supposed to work (= be checked)? -->
<!-- Try loading a marked-as-malware css through an @import from a clean URI -->
<link rel="stylesheet" type="text/css" href="import.css"></link>
</head>

<body onload="checkLoads()">
The following should not be hidden:
<div id="styleCheck">STYLE TEST</div>
</body>
</html>
